Remove numpy and parsesetup dependencies
Numpy is not used, but is required because of parsesetup which is not
properly packaged.
This commit removes the use for numpy and parsesetup by replicating the
needed behavior from parsesetup.
The replacement function can be improve me by mocking failing imports in
the setup.py file.
In the meantime, if setup.py includes other imports that perform
computation over the parameters that are passed to "setup(...)", this function or any library relying on monkey patching of "setup(...)" (as parsesetup) will give bad results.